DIY: Fixed compiler warnings

+++ external/OAuthConsumer/NSMutableURLRequest+Parameters.m
@@ -66,7 +66,7 @@
 {
     NSMutableString *encodedParameterPairs = [NSMutableString stringWithCapacity:256];
     
-    int position = 1;
+    NSUInteger position = 1;
     for (OARequestParameter *requestParameter in parameters){
         [encodedParameterPairs appendString:[requestParameter URLEncodedNameValuePair]];
         if (position < [parameters count])
